Initial cleaning of e-matching instantiation strategy

+++ src/theory/quantifiers/ematching/inst_strategy_e_matching.h
@@ -47,9 +47,9 @@ public:
   /** add pattern */
   void addUserPattern( Node q, Node pat );
   /** get num patterns */
-  int getNumUserGenerators( Node q ) { return (int)d_user_gen[q].size(); }
+  size_t getNumUserGenerators( Node q ) { return d_user_gen[q].size(); }
   /** get user pattern */
-  inst::Trigger* getUserGenerator( Node q, int i ) { return d_user_gen[q][ i ]; }
+  inst::Trigger* getUserGenerator( Node q, size_t i ) { return d_user_gen[q][ i ]; }
   /** identify */
   std::string identify() const override { return std::string("UserPatterns"); }
 };/* class InstStrategyUserPatterns */
